requirements.txt
Django==4.2.8
asgiref==3.7.2
crispy-bootstrap5==2023.10
django-crispy-forms==2.1
pip==22.3.1
setuptools==65.5.1
sqlparse==0.4.4
typing-extensions==4.8.0
wheel==0.38.4
djangorestframework~=3.14.0
djangorestframework-simplejwt==5.3.1